Octopus: Enable PE5 as a precaution
Setting PE5 to high makes sure the hotend fan is running to avoid heat-creep. The fan is turned off automatically once Klippy has started on the pi and communicates with the MCU. There is a matching PR in the Voron-2 repo to ensure the same thing in case of a MCU shutdown.