Add a beginner's guide to Klipper macros
There seem to be a lot of questions on the Voron Discord about writing
Klipper macros. Most of them from users who are not familiar with the
internals of Klipper and macro writing, asking simple questions.

This guide is an attempt to describe how to begin to understand and
write Klipper macros.