Add a beginner's guide to Klipper macros (#305)
* Add a beginner's guide to Klipper macros

There seem to be a lot of questions on the Voron Discord about writing
Klipper macros. Most of them from users who are not familiar with the
internals of Klipper and macro writing, asking simple questions.

This guide is an attempt to describe how to begin to understand and
write Klipper macros.

* Add content documention save/restore GCode state

Add some text describing Klipper's GCode state and how to
save and restore it.

* Add content to Klipper macro guide for slicer integration

Add some content to the Klipper macro guide that explains how to
pass parameters to macros from slicers.

* Fix guide name spelling mistake

Rename the guide to "Klipper_Macros_Beginners_Guide.md" to fix
a spelling mistake in the name. Also, fix links using the wrong
name.

* Provide an example for Klipper's macro evaluation

The way macros templates are evaluated by Klipper needed an
example since the concept might not be easy to understand.

Provide that example.

Co-authored-by: Mitko Haralanov <dharalanov@nvidia.com>